[cleanup] Remove interwiki_forwarded_from list from family files
interwiki_forwarded_from was implemented in compat but that feature wasn't
merged into the already existing core titletranslate.py. As a result,
interwiki_forwarded_from is unimplemented in core.

titletranslate is used by interwiki.py only which is used very rare
(is it still used?). Nobody missed this feature for so many years
and it can be removed then.

Bug: T104125
Change-Id: I5bd27becb735c2ecd817a3c27490dfeeca0fa948