Use ignore_discard for login cookie container
Allow saving and loading cookies with the discard flag
(ie. session cookies). This allows saving bot password sessions,
which did not work before (normal sessions did since they can fall
back on the user_token cookie which is non-session). Since session
cookies do not have an expiry, the login will be valid as long as
the session exists on the server (e.g. for Wikimedia servers this
is nowadays 24 hours).

Bug: T261066
Change-Id: I1519fd7cde847ab82f378758e04024619410922f