Revert "Always return raw generators from APISite methods, not objects"
This reverts commit 8ec067010722940f121ce4b97890f39f3d15df18.

Reason for revert: The main reason is that pagegenerators does not work with 
this change. getCombinedGenerator enables namespaces filtering. 
NamespaceFilterPageGenerator loads all items even they are not needed which may
need a lot of time. Site.allpages will always fail if no namespaces is given to
the generator. You cannot specify a page generator in the current way for that
patch

Bug: T263916
Change-Id: I459102e9a0f1e0672d1a54eeb33b861c04201d51