[IMPR] Create a SiteLink with __getitem__ method
SiteLinkCollection initializer updates data contents to self._data
dict. The update method calls __setitem__ for each data entry; for
each entry a SiteLink is created which needs a siteinfo content for
every affected site and leads to a huge loading time.

Now with this patch the SiteLink object is created when the link
is fetched from self._data via __getitem__ or __iter__ method.
This prevents siteinfo content for each site from loading even
a link is currently not used and decrease loading time a lot.

Bug: T245809
Bug: T226157
Bug: T238471
Bug: T273386
Change-Id: Icc92cfd5d200fde47eb127e2fe601f812427624a