[IMPR] collect family files from zip folder
if the given folder_path of register_family_file funtion is not a
directory it probably resides in a zip folder. In that case check
the current path and its parents whether it is a zip file. If no
zipfile is found raise a NotADirectoryError. Otherwise collect all
files inside the folder and filter the family files by the suffix.

Bug: T278076
Change-Id: Ifb4a75c8f81a7d727b149ed6354e8896ffeb4e0c