[cleanup] Rename is_IP to is_ip_address
My hat's off for already migrating to ipaddress [1]. I didn't realize Python
added that, and like you implemented my own helpers for this [2][3]. Neat
that this is no longer necessary.

I think we should drop it but Xqt disagrees, so just cleaning up the tests and
renaming it to is_ip_address ('ip' stands for 'internet protocol' so the prior
'is_IP' name was vague).

[1] https://phabricator.wikimedia.org/T76286
[2] https://gitweb.torproject.org/stem.git/tree/stem/util/connection.py#n433
[3] https://gitweb.torproject.org/stem.git/tree/test/unit/util/connection.py#n460

Bug: T281200
Change-Id: I7ed46e885c608d34c6046921ba1a45b5e024406a