[IMPR] Remove distutils
Python 3.10 deprecated distutils and it will be removed in 3.12. They advise
that we use packaging instead...

  https://docs.python.org/3.10/whatsnew/3.10.html#distutils-deprecated

We mostly used distutils for version parsing, for which packaging provides
a simple replacement...

  https://stackoverflow.com/a/11887885
  https://packaging.pypa.io/en/latest/version.html#packaging.version.parse

Setuptools contains a copy of packaging under 'pkg_resources.extern.packaging'
and pkg_resources.parse_version() is effectively an alias of
packaging.version.parse(). Since we already have a depency on Setuptools
this uses its copy.

As for get_python_lib() sysconfig can provide the same...

  >>> import distutils.sysconfig
  >>> distutils.sysconfig.get_python_lib(standard_lib=True)
  '/usr/lib/python3.8'

  >>> import sysconfig
  >>> sysconfig.get_paths()['stdlib']
  '/usr/lib/python3.8'

Bug: T274696
Change-Id: I6245895d4249d32cfae34043eed489cc94ddea2e