[W0120] Useless else clause detected on the loop
When a loop specifies no break statement, the else clause will always
execute when the loop sequence is empty, thus making it useless. It is
recommended to have the statements under else in the same scope as the
loop itself. If the else clause does not always execute at the end of a
loop clause, then the code should add a break statement within the loop
block.

Change-Id: I57a8b70da399a869554c8ab78e6eefad18ac1f6d