imagetransfer.py: allow images existing in the shared repo
When transferring files from Commons to other wikis that use
Commons as a shared repo, they can't be uploaded, as they are
seen as a duplicate of the image in the shared repo.

Add a parameter -force_if_shared to override this check.

Note, the user still needs the 'reupload-shared' right at the
target wiki.

Bug: T267535
Change-Id: Ia4412c84aef09d2353017853e327071fffb07386