Drop piprop from meta=proofreadinfo API call
Because the "pi" prefix is already in use by PageImages, this module
is being changed to use a prefix of prpi instead. This should not be
a breaking change in either direction because the default value of
the parameter included all of the current prop values.

Because the default parameter for piprop was always
'namespaces|qualitylevels', it is safe to remove this parameter
entriely, thus avoiding an unrecognized parameter warning.

Bug: T290585
Change-Id: I0f269e06a10c75c296d9c126228970b7ccff72d4