Avoid non-deteministic behavior in removeDisableParts
The order of iteration over a set can change everytime a new
Python process is run because the string hashes are different.[1]
This means the function can apply replacements in different order
which may result in different output. So don't cast the sequence
to a set and use any sets at all, it isn't really necessary.

Add a regression test.

[1] https://docs.python.org/3/reference/datamodel.html#object.__hash__

Change-Id: If66173a7301d308b7addd7f63c7a1d2a2771abbc