[IMPR] Implement a new and easier color format for Pywikibot
The old color format is like 'this is a \03{green}colored\03{default} text'.
With Python 3 curly brackets are parts of the default format string. Using
it needs strings like
  'this is a \03{{green}}{color}colored\03{{default}} text'
  .format(color='red')
which needs double brackets.

A color_format method was introduced to omit the \03 escape sequence which
allows strings like
  color_format('this is a {green}{col} colored{default} text', col='red')
but be aware the 'color' key is reserved for a color key and will be
escaped then. In short: it is not very helpful to simplify colored text
because each string must be passed through that function.

Now a new color pattern is implemented which can be used directly with output
methods and does not bother any format string. The sample above can be
written like

  'this is a <<green>>{color} colored<<default>> text'.format(color='red')

The advantages are
- the escape sequence \03{...} is replaced by <<...>>
- the color_format method will be deprecated, there is no benefit to
  keep it.
- the old and new implementation will be supported both for a while
  but must not be mixed in the same output statement
- a usage sample is made in pwb.py

Change-Id: I5bfcaac8ef12a65dab1a2515809373633e1706b1