[cleanup] remove userinterfaces.gui.ListBoxWindow
remove ListBoxWindow which was introduced in compat 2004 but was neither
used anywhere in the framework nor is it documented how to use it.
After creating the listbox window the only choices is to close the window
by clicking ok. There is no method tho get an element of the numbered list.

Change-Id: Ief1a3172bdacbb4c45a66d834ce31247d3a64a82