[IMPR] Assert internal state on Claim modifications
Add an assertion to ensure some operations cannot be
performed on a qualifier or a reference (such as
adding qualifiers/references or changing the rank)
and another one to ensure some operations can only
be performed on an existing (saved) claim.

Note that in the second case the operations would
already crash. But now the error messages are more
clear.

Change-Id: If7867b67ac0af72a5be0650ece2813fe1a964077