[bugfix] Fix setup.read_desc() regex
If a space was behind a sphinx link and the link was different from
the label, the link was removed but the alignment does not match the
previous line. Modify the regex pattern to remove a following space
in such case.

Also test the long_description created by setup.py in make_dist.py
script

Bug: T327056
Change-Id: Ie71c4f9feb38e4ba822330b6978d19f506304c95