[Bugfix] Copy snak IDs/hashes when using editEntity
Previously, any new claims added manually to the item via direct access
to `item.claims` would keep on getting added again every time editEntity
was called. By assigning the snak IDs/hashes that are created
when a new claim/qualifier/reference is created,
duplicates should not occur.

Note: To allow flake8 to pass, I had to reduce the names of variables.
The original names can be found as a comment on the patchset.

Bug: T327607
Change-Id: I26d043702d7cef63afcfc103ce7206ca9c8e5c41