[bugfix] Take into account that a repository clone may have no tags
Pywikibot repository may be cloned without depth and may have no tags.
This patch solves the IndexError which is raised in such case. This is
necessary for CI tests.

Call sys.exit(1) if an error occurs to signal CI tools to fail the test.
Also update coverage and use f-strings

Bug: T335676
Change-Id: I9d8b2990e286772efe4827c2a11fee50af598308