[cleanup] Remove unused _query_continue method.
QueryGenerator._query_continue method is no longer used since
b8060a7e8b8553f5beca6046809c7f99cc66a0c8 and can be removed.

rename _continue() method to continue_update and merge it
with _add_continues. Always return None with it instead of
False because the result is not important anymore.
This looks like a breaking change but it does not care.

Bug: T343204
Change-Id: I4eaa37f150f8c0d9a008419637b7c320385708b0