Wrap DOT-string in curly braces in category_graph.py
Invalid DOT-string in category_graph.py caused TypeError exception when
running category_graph.py. Wrapped the DOT-string in curly braces in
argument passed to pydot.graph_from_dot_data() so that the DOT-string is
in line with the pydot documentation.

Wrap DOT-string in additional set of curly braces.

I forgot the second set of curly braces in the original commit and am
adding them in this commit. The second braces escape the first set of
braces and resolve "style" in the argument.

Bug: T346007
Change-Id: I635e077240b7d6441d5349edf261ba00a96748e7