[doc] Center branding text in sidebar
Most recent documentation builds no longer center the text in the
sidebar. Instead, it's aligned to the left. This is probably due
to changes in Sphinx 6.0.0 that broke a variable Furo relies on
to center that text. I was not able to pinpoint the root cause so
I'm proposing custom CSS as a workaround. Tested with Sphinx 7.2.3
and Furo 2023.09.10.

Change-Id: Ib33c97355cfe2b696d72c5520bcd69a2c691bfec