Validate file path input for category_graph.py.
Validate file path input for category_graph.py in order to handle
nonexistent paths or undesired overwrites BEFORE the script runs. On
nonexistent path, path will be created. If file exists, user will be
given y/n prompt to override with "n" exiting bot and any other key
overwriting. Switched from input() to pywikibot.input_choice().

Update logic for file path creation.

Because the path specified by -to is not an actual file in that it has
no file extension, special logic is required in order to create the
directory appropriately. If the path does not contain '/' it is assumed
to be at the top level, and no directory creation is necessary.
Otherwise, the directory is created if it does not exist.

Change to non-reserved word for input result.

Change from "prompt" to "choice" for prompt variable name, as "prompt"
is a part of the standard lib. Minor syntax changes to follow
guidelines.

Change to double quotes to single quotes.`

Replace reserved word 'input'. Use pathlib to extract folder and file.

The reserved word 'input' was still being used. This has been replaced
by the word 'choice'. Pathlib is now being used to extract the folder
and file in order to gracefully handle filepaths containing '/' or '\'.

Revise exception handling. Use path.mkdir instead of os.mkdir.

Bug: T346417
Change-Id: I637bc0ef0bc99d5a59095890cc6b5be99bfd34fb