[FEAT] Add a filter to check if a page is redirect
Added a -redirect filter to assert if page is a redirect. Likewise,
the filter can be negated by using -redirect:false.

This is a missing feature present in compat, which implemented -redirectonly,
but this is more useful because it can be negated, and the name is different
to make it clearer.

Bug: T261549
Change-Id: If1a2c9bb420d9b5af69371537c7ca724096d8a23