[fix] check for valid family and site option after -help is processed
pywikibot.handle_args() checks for a valid default site and prints an
error if either the family or the site code is invalid.

Make this test after -help option is processed to avoid unnecessary
api call when -help option is used. This is a workaround unless the
APISite.login(cookie_only=True) call no longer retrieves userinfo in
an early state i.e. instantiating the site.

Bug: T350272
Change-Id: I04a55c96a826c76b5bcf89893be92413985c6429