[cleanup] Make Family.languages_by_size dynamic
Currently languages_by_size data is obtained from
https://wikistats.wmcloud.org/ via wikimedia_sites script which
re-writes the family files for familes which have data in the
wikistats database.

The languages_by_size sorting codes are only used by interwiki.py
script with -hint option together with a number.

This change replaces the current system with extraction of the
statistics from wikistats of each site in the family, done on demand.

- change language_by_size attributes to sorted codes
- add language_by_size class property to WikimediaFamily family
- remove generic languages_by_size and codes
- use Family.codes instead of Family.languages_by_size if appropriate
- use a global dict on wikistats to hold the cache
- remove wikimedia_sites maintenance script which is no longer needed
- add family_list to unidata.py
- update tests
- update documentation

Bug: T78396
Change-Id: Idb7eb45ec2db002f13c62c02156dd3d7d97e1e04