[IMPR] remove sorting from page_gen generator
The page_gen method retrieves pages by the get_page() index which
lies within start and end parameters. The sorting key is found by
get_number() function which gives the same value which is previously
given by the index of get_page. Therefore the sorting is needless but
the sorting build-in function is greedy and retrieves all pages which
is not in the sense of the generator.

Change-Id: Ibbcf55f02500fd1c0bd4e95bcac267aded4fc777