[fix] check whether tile exists first
This is a temporary solution for T372106.
FilePage.exists() gives True if the file decription exists even there
is no file at all but file_is_shared() fails with missing imageinfo
exception. We should have a way to check for existing image.

For this special case such files should be delinked.

Bug: T372106
Change-Id: Ieec85c7de50ba65433f2d9731aaf65316fb74c55