Cleanup: Remove Structured Discussions/Flow support
Structured Discussion (previously known as Flow) support was introduced
to Pywikibot with release 3.0.20170403. But this extension is no longer
under active maintenance for 7 years and people are encouraged to no
longer use it. On most wikis structured discussions are in read-only
mode and archived as subpages. The remaining wikis cawikiquote,
fiwikimedia, gomwiki, kabwiki and ptwikibooks will be processed shortly.

Structured Discussions/Flow support is deprecated since Pywikibot 9.4.

Bug: T381551
Change-Id: I2443b48e1cf3ff77813d9ad8ddfedb1b01d946d0