[IMPR] Add return after super(UI).encounter_color() call
If none of stdin, stdout, or stderr is used, Win32UI.encounter_color()
calls the base class implementation. This is not critical because the
base class only raises NotImplementedError. However, returning after
this call is safer, since the base implementation might change in the
future and 'addr' would not be defined in such a case, leading to a
NameError.

Change-Id: I06007e07acfe2a64c564eaa553e4e897cbcf770d